rag/requirements.txt
anthropic>=0.49.0
openai>=1.50.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pypdf>=4.3.0
tiktoken>=0.8.0
python-dotenv>=1.0.0
